+++ app/assets/stylesheets/default.css
@@ -2,7 +2,7 @@
 body {
 	font-family: Helvetica, Arial, Geneva, sans-serif;
 	padding: 2em;
-	border: 8px solid teal;
+    border: 8px solid teal;
 }
 
 span.word {
@@ -14,29 +14,7 @@ span.word {
 	padding-left: 2em;
 }
 
-span.guesses {
-	color: #ff2600;
-	font-weight: bold;
-	letter-spacing: 4px;
-	text-transform: uppercase;
-	padding-left: 1em;
-}
-
-span.lose {
-	color: red;
-	font-weight: bold;
-}
-
-span.error {
-  color: red;
-}
-
-span.win {
-        color: green;
-        font-weight: bold;
-}
-
 form {
-	border: 1px solid grey;
+	/* border: 1px solid grey; */
 	margin: 4px;
 }
