--- app/models/hangperson_game.rb
@@ -1,48 +1,0 @@
-class HangpersonGame
-
-  attr_accessor :word, :guesses, :wrong_guesses
-
-  def initialize(word)
-    @word = word
-    @guesses = ''
-    @wrong_guesses = ''
-  end
-
-  def guess(letter)
-    if letter.nil? || letter !~ /\A[a-zA-Z]\Z/i # match single letter,ignoring case
-      raise ArgumentError
-    end
-    letter = letter[0].downcase
-    if  @word.include?(letter) && !@guesses.include?(letter)
-      @guesses += letter
-    elsif !@word.include?(letter) && !@wrong_guesses.include?(letter)
-      @wrong_guesses += letter
-    else
-      false
-    end
-  end
-
-  def word_with_guesses
-    @word.gsub(/./)  { |letter| @guesses.include?(letter) ? letter : '-' }
-  end
-
-  def check_win_or_lose
-    if word_with_guesses == @word
-      :win
-    elsif @wrong_guesses.length > 6
-      :lose
-    else
-      :play
-    end
-  end
-
-  # Get a word from remote "random word" service
-
-  def self.get_random_word
-    require 'uri'
-    require 'net/http'
-    uri = URI('http://watchout4snakes.com/wo4snakes/Random/RandomWord')
-    Net::HTTP.post_form(uri ,{}).body
-  end
-
-end
