--- app/views/game/lose.html.erb
@@ -1,10 +1,0 @@
-<h1>Sorry, you lose!</h1>
-
-<p>
-   The word was
-   <span class="lose"><%= @game.word %></span>
-</p>
-
-<%= render :template => 'game/new' %>
-
-
