--- app/views/game/win.html.erb
@@ -1,9 +1,0 @@
-<h1>You Win!</h1>
-
-<p>
-  The word was
-  <span class="win"><%= @game.word %></span>.
-</p>
-
-<%= render :template => 'game/new' %>
-
