+++ app/views/layouts/application.html.erb
@@ -6,6 +6,7 @@
     <%= csrf_meta_tags %>
   </head>
   <body>
+    <%= button_to "Home Page", { :action => "index" }, :method => :get %>
     <div class="container">
       <div id="header" class="span-24">
         <h>Girl Scout Cookies Exercise Tracker</h>
