+++ app/views/measurements/enter_my_measurements.html.erb
@@ -8,4 +8,4 @@
   <%= submit_tag 'Enter' %>
 <% end %>
 
-<%= render :template => 'game/new' %>
+<%= render :template => 'measurements/index' %>
