+++ app/views/measurements/index.html.erb
